pyproject.toml
[tool.poetry]
name = "video-captioning-bot"
version = "0.1.0"
description = "A Python bot for adding captions and voiceovers to videos"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "*.py"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
moviepy = "^2.1.2"
openai = "^1.0.0"
gtts = "^2.3.2"
python-dotenv = "^1.0.0"
requests = "^2.28.0"
numpy = "^1.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"